Separate user script testing fro NotebookInstallSpec
