+++ automation/src/test/scala/org/broadinstitute/dsde/workbench/leonardo/NotebookInstallSpec.scala
@@ -30,42 +30,15 @@ class NotebookInstallSpec extends ClusterFixtureSpec {
       }
     }
 
-    // requires a new cluster because we want to pass in a user script in the cluster request
-    "should allow user to create a cluster with a script" taggedAs Tags.SmokeTest in { clusterFixture =>
+    // This is the negative counterpart of the NotebookUserScriptSpec test named
+    // "should allow importing a package that requires a user script that IS installed"
+    // to verify that we can only import 'arrow' with the user script specified in that test
+    "should error importing a package that requires an uninstalled user script" in { clusterFixture =>
       withWebDriver { implicit driver =>
         //a cluster without the user script should not be able to import the arrow library
         withNewNotebook(clusterFixture.cluster) { notebookPage =>
           notebookPage.executeCell("""print 'Hello Notebook!'""") shouldBe Some("Hello Notebook!")
           notebookPage.executeCell("""import arrow""").get should include("ImportError: No module named arrow")
-        }
-      }
-
-      // create a new bucket, add the user script to the bucket, create a new cluster using the URI of the user script and create a notebook that will check if the user script ran
-      val gpAllocScriptProject = claimGPAllocProject(hermioneCreds)
-      val billingScriptProject = GoogleProject(gpAllocScriptProject.projectName)
-      withNewGoogleBucket(billingScriptProject) { bucketName =>
-        val ronPetServiceAccount = Sam.user.petServiceAccountEmail(clusterFixture.billingProject.value)(ronAuthToken)
-        googleStorageDAO.setBucketAccessControl(bucketName, EmailGcsEntity(GcsEntityTypes.User, ronPetServiceAccount), GcsRoles.Owner)
-
-        val userScriptString = "#!/usr/bin/env bash\n\npip2 install arrow"
-        val userScriptObjectName = GcsObjectName("user-script.sh")
-        val userScriptUri = s"gs://${bucketName.value}/${userScriptObjectName.value}"
-
-        withNewBucketObject(bucketName, userScriptObjectName, userScriptString, "text/plain") { objectName =>
-          googleStorageDAO.setObjectAccessControl(bucketName, objectName, EmailGcsEntity(GcsEntityTypes.User, ronPetServiceAccount), GcsRoles.Owner)
-          val clusterName = ClusterName("user-script-cluster" + makeRandomId())
-
-          withWebDriver { implicit driver =>
-            withNewCluster(clusterFixture.billingProject, clusterName, ClusterRequest(Map(), None, Option(userScriptUri)), apiVersion = V2) { cluster =>
-              Thread.sleep(10000)
-              withNewNotebook(cluster) { notebookPage =>
-                notebookPage.executeCell("""print 'Hello Notebook!'""") shouldBe Some("Hello Notebook!")
-                notebookPage.executeCell("""import arrow""")
-                notebookPage.executeCell("""arrow.get(727070400)""") shouldBe Some("<Arrow [1993-01-15T04:00:00+00:00]>")
-              }
-            }(ronAuthToken)
-          }
-
         }
       }
     }
